requirements.txt
asgiref==3.3.1
autopep8==1.5.4
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coverage==5.3.1
cryptography==36.0.0
Django==3.1.4
django-countries==7.2.1
flake8==3.8.4
flake8-django==1.1.1
gunicorn==20.1.0
idna==3.3
invoke==1.6.0
isort==5.7.0
mccabe==0.6.1
paramiko==2.8.1
pathlib2==2.3.6
Pillow==8.0.1
psycopg2-binary==2.9.2
pycodestyle==2.6.0
pycparser==2.21
pyflakes==2.2.0
PyNaCl==1.4.0
python-dotenv==0.19.2
pytz==2020.4
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
stripe==2.63.0
testfixtures==6.17.1
toml==0.10.2
urllib3==1.26.7
